외부 3D 모델 I/O 작업 완료
