--- learnMetal_2/Renderer.swift
@@ -1,14 +1,0 @@
-//
-//  Renderer.swift
-//  learnMetal_2
-//
-//  Created by Hyeok Cho on 9/8/25.
-//
-
-//TODO: 코드 분리
-import MetalKit
-
-class Renderer {
-    
-    func renderFrame() { }
-}
