+++ learnMetal_2/Ttouch.mtl
@@ -1,0 +1,17 @@
+# Blender 4.5.0 MTL File: 'TtouchMultiAnim_1.blend'
+# www.blender.org
+
+newmtl Ttouch
+Ks 0.500000 0.500000 0.500000
+Ke 0.000000 0.000000 0.000000
+Ni 1.500000
+d 1.000000
+illum 2
+Pr 0.500000
+Pm 0.000000
+Ps 0.000000
+Pc 0.000000
+Pcr 0.030000
+aniso 0.000000
+anisor 0.000000
+map_Kd Texture_01.png
